moved the dose rate tensor processing to the general utility script
